backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pyjwt==2.8.0
gitpython==3.1.42